backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
openai==1.12.0
playwright==1.41.2
celery==5.3.6
redis==5.0.1
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
boto3==1.34.34
python-multipart==0.0.6
httpx==0.26.0
pytest==8.0.0
pytest-asyncio==0.23.5
alembic==1.13.1

passlib[bcrypt]==1.7.4
PyJWT==2.11.0
bcrypt==3.2.2
